Make MediaStreamTrackPrivate RefCounted https://bugs.webkit.org/show_bug.cgi?id=201040
Reviewed by Darin Adler.

MediaStreamTrackPrivate should not be ref/deref except from the main thread.
The only method called from a background thread is audioSamplesAvailable in which
it is unsafe to ref a MediaStreamTrackPrivate.
Make the track RefCounted will ensure that no ref/deref is made in background threads.
No observable change of behavior.

* platform/mediastream/MediaStreamTrackPrivate.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249134 268f45cc-cd09-0410-ab3c-d52691b4dbfc