Remove support for macOS < 10.13 (part 2) https://bugs.webkit.org/show_bug.cgi?id=201197 <rdar://problem/54759985>
Source/JavaScriptCore:

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

Reviewed by Darin Adler.

* Configurations/FeatureDefines.xcconfig:

Source/WebCore:

Reviewed by Darin Adler.

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

No new tests -- no new or changed functionality.

* Configurations/FeatureDefines.xcconfig:

Source/WebCore/PAL:

Reviewed by Darin Adler.

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

* Configurations/FeatureDefines.xcconfig:

Source/WebKit:

Reviewed by Darin Adler.

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

* Configurations/FeatureDefines.xcconfig:
* Configurations/WebKit.xcconfig:

Source/WebKitLegacy/mac:

Reviewed by Darin Adler.

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

* Configurations/FeatureDefines.xcconfig:

Tools:

Reviewed by Darin Adler.

Update conditionals that reference WK_MACOS_1013 and suffixes like
_MACOS_SINCE_1013, assuming that we're always building on 10.13 or
later and that these conditionals are always True or False.

See Bug 200694 for earlier changes in this area.

* TestWebKitAPI/Configurations/FeatureDefines.xcconfig:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249208 268f45cc-cd09-0410-ab3c-d52691b4dbfc