AccessibilityRenderObject::setSelectedTextRange fails to set the selection passed an empty line. https://bugs.webkit.org/show_bug.cgi?id=201518 <rdar://problem/54835122>
Patch by Andres Gonzalez <andresg_22@apple.com> on 2019-09-06
Reviewed by Ryosuke Niwa.

Source/WebCore:

Test: accessibility/set-selected-text-range-after-newline.html

In the case of an empty line, the CharacterIterator range start and end
were not equal, thus we were not advancing the iterator and returning
the iterator range end, which is not correct. With this change we are
always advancing the iterator if its text is just '\n'. This covers all
the cases we fixed before plus empty lines.

* editing/Editing.cpp:
(WebCore::visiblePositionForIndexUsingCharacterIterator):

LayoutTests:

Extended this test to set the selection range passed an empty line.
* accessibility/set-selected-text-range-after-newline-expected.txt:
* accessibility/set-selected-text-range-after-newline.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249565 268f45cc-cd09-0410-ab3c-d52691b4dbfc