Stop using testRunner.setPrivateBrowsingEnabled_DEPRECATED() in http/tests/workers/service/basic-register.html https://bugs.webkit.org/show_bug.cgi?id=201590
Reviewed by Youenn Fablet.

Tools:

Add a new API test to check that if the default session registers a service worker, the
hasServiceWorkerRegistration flag is still false in an ephemeral session.

* TestWebKitAPI/Tests/WebKitCocoa/ServiceWorkerBasic.mm:

LayoutTests:

Split the layout test in 2 to test both the default and the ephemeral session without relying on the
testRunner.setPrivateBrowsingEnabled_DEPRECATED() hack.

* http/tests/workers/service/basic-register-expected.txt:
* http/tests/workers/service/basic-register-private-expected.txt: Copied from LayoutTests/http/tests/workers/service/basic-register-expected.txt.
* http/tests/workers/service/basic-register-private.html: Added.
* http/tests/workers/service/resources/basic-register.js:
(async.test):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249648 268f45cc-cd09-0410-ab3c-d52691b4dbfc