Unreviewed, fixed a crash in ssl/applepay/ApplePayCancelEvent.https.html
After r250060 (which landed very shortly after the patch that introduced ApplePayCancelEvent),
all Events must be IsoHeap-ed. ApplePayCancelEvent was not IsoHeap-ed, and so this caused
assertion failures in ssl/applepay/ApplePayCancelEvent.https.html.

Added the WTF_MAKE_ISO_ALLOCATED macros to ApplePayCancelEvent without a review to stop the
test from failing in EWS.

* Modules/applepay/ApplePayCancelEvent.cpp:
* Modules/applepay/ApplePayCancelEvent.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250080 268f45cc-cd09-0410-ab3c-d52691b4dbfc