Web Inspector: Improve the Uncaught Exception View file a bug link https://bugs.webkit.org/show_bug.cgi?id=201717
Reviewed by Devin Rousso.

Source/WebInspectorUI:

* UserInterface/Debug/UncaughtExceptionReporter.js:
Allow the link to be clicked. Use openInNewTab on click to also
bring the new tab to the foreground. Also update the content.

Source/WebKit:

* UIProcess/WebInspectorProxy.cpp:
(WebKit::WebInspectorProxy::bringInspectedPageToFront):
* UIProcess/WebInspectorProxy.h:
* UIProcess/WebInspectorProxy.messages.in:
Provide a way to bring the inspected page to the foreground.

* WebProcess/WebPage/WebInspectorUI.cpp:
(WebKit::WebInspectorUI::openInNewTab):
Use it when opening a new tab beside the inspected page.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250243 268f45cc-cd09-0410-ab3c-d52691b4dbfc