IDBRequest should not prevent a page from entering the back/forward cache https://bugs.webkit.org/show_bug.cgi?id=202270 <rdar://problem/55744717>
Reviewed by Youenn Fablet.

Source/WebCore:

Update IDBRequest::canSuspendForDocumentSuspension() to return true instead of
false. IDBRequest can only fire 2 events (error / success) and for both its
merely enqueues the Event to the DocumentEventQueue. The DocumentEventQueue
uses a SuspendableTimer (which will be suspended while the document is in the
page cache) so there is no risk of firing the events (and thus running script)
while in the page cache.

Test: storage/indexeddb/IDBRequest-page-cache.html

* Modules/indexeddb/IDBRequest.cpp:
(WebCore::IDBRequest::canSuspendForDocumentSuspension const):

LayoutTests:

Add layout test coverage.

* storage/indexeddb/IDBRequest-page-cache-expected.txt: Added.
* storage/indexeddb/IDBRequest-page-cache.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250425 268f45cc-cd09-0410-ab3c-d52691b4dbfc