[Nicosia] Enable runtime checks and triggers for async non-main-frame scrolling https://bugs.webkit.org/show_bug.cgi?id=202509
Reviewed by Carlos Garcia Campos.

Source/WebCore:

Add build guards that would enable runtime checks of the asynchronous
frame scrolling setting in the ScrollingCoordinator and
RenderLayerCompositor classes for the Nicosia-using ports.

The check in ScrollingCoordinator avoids bailing on coordinating
scrolling for non-main frames when the setting is enabled. The check in
RenderLayerCompositor avoids imposing a requirement for a composition
layer for any non-main frame when the setting is disabled.

* page/scrolling/ScrollingCoordinator.cpp:
(WebCore::ScrollingCoordinator::coordinatesScrollingForFrameView const):
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForScrollableFrame const):

Source/WebKit:

* WebProcess/WebCoreSupport/WebChromeClient.h:
Allow the non-main-frame compositing trigger for the Nicosia-using ports
as well. This will enable non-main-frame compositing if the related
setting is also enabled.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250650 268f45cc-cd09-0410-ab3c-d52691b4dbfc