REGRESSION(251087): Several API tests are failing https://bugs.webkit.org/show_bug.cgi?id=202961
Reviewed by Brent Fulgham.

WebPageProxy::maybeInitializeSandboxExtensionHandle returns early if the WebContent process is currently launching,
since the audit token is needed to create the sandbox extension. The audit token will not be available until the
process has finished launching. In the same way, the method should return early if the WebContent process has
terminated.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::maybeInitializeSandboxExtensionHandle):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251116 268f45cc-cd09-0410-ab3c-d52691b4dbfc