Invalid instruction generated for ARM_THUMB2 in llint https://bugs.webkit.org/show_bug.cgi?id=202844
Reviewed by Saam Barati.

Do not allow instruction execution to reach OSR return label on ARMv7.
Currently we are seeing the instruction execution hitting a .word directive in
the instruction stream and segfaulting. There are two words used to represent a
global label which was generated as part of the work on OSR Exit to LLInt work
(r250806). The double word generation only occurs in ARMv7 and therefore only here
the segfault manifests itself.

* llint/LowLevelInterpreter.asm:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251196 268f45cc-cd09-0410-ab3c-d52691b4dbfc