Web Inspector: Elements: selection shouldn't be dimmed by shadow trees https://bugs.webkit.org/show_bug.cgi?id=203057
Reviewed by Joseph Pecoraro.

* UserInterface/Views/DOMTreeOutline.css:
(.tree-outline.dom li .selection-area):
(.tree-outline.dom li > span): Added.
Don't force the `.selection-area` to be behind (`z-index: -1;`) the rest of the content, as
that also puts it behind any `background-color`. Instead, we should make all the content in
every `WI.DOMTreeElement` be above (`z-index: 1;`) the rest of the content, including the
`.selection-area` and any parent's `background-color`.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251254 268f45cc-cd09-0410-ab3c-d52691b4dbfc