[MSE][GStreamer] Reintroduce ensureGLVideoSinkContext() fix https://bugs.webkit.org/show_bug.cgi?id=201401
Reviewed by Philippe Normand.

That patch was almost mechanically reverted temporarily along the rest
of the WebKitMediaSrc rework. In theory that should have been OK since
it was supposed to fix a regression of that patch... But actually
the bug was already causing problems before, the WebKitMediaSrc rework
just exposed a more common case.

It has been reported that the same bug arises when adding some CSS
styles to the <video> element, and the patch fixes the issue.

And the patch actually makes sense in any case... since without it
some perfectly valid state transitions are not covered!

So I'm reintroducing it.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::changePipelineState):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251544 268f45cc-cd09-0410-ab3c-d52691b4dbfc