REGRESSION (r251261?): Flaky Test: inspector/runtime/getProperties.html https://bugs.webkit.org/show_bug.cgi?id=203271 <rdar://problem/56557216>
Reviewed by Devin Rousso.

r251261 was reverted and the test has become not so flaky, but the previous flakiness showed us a potential bug
in the test: it is possible that the console message for unhandled rejected promise gets dispatched after
dumping the results. To make the test more robust, stop showing that message in console.

* inspector/runtime/getProperties-expected.txt:
* inspector/runtime/getProperties.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251618 268f45cc-cd09-0410-ab3c-d52691b4dbfc