REGRESSION(2.27.2): [GTK] Overview item remove button broken https://bugs.webkit.org/show_bug.cgi?id=203461
Reviewed by Adrian Perez de Castro.

This is a regression of the switch to new custom protocols implementation. Epiphany is passing NULL as content
type of ephy-resource requests, in which case we should try to guess the mime type, but we are not doing it.

* UIProcess/API/glib/WebKitURISchemeRequest.cpp:
(webkitURISchemeRequestReadCallback): Use MIMETypeRegistry to guess the mime type when content type is nullptr,
like NetworkDataTask does.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251769 268f45cc-cd09-0410-ab3c-d52691b4dbfc