[iOS] imported/mozilla/svg/text/textpath-selection.svg is still flaky https://bugs.webkit.org/show_bug.cgi?id=203659 <rdar://problem/52124292>
Reviewed by Tim Horton.

This test still fails in some internal test runner configurations, because the iOS text selection grabber dots
sometimes show up in the actual result and not in the expectation. This still happens even after suppressing
UITextSelectionView during ref test snapshotting in r251526, since the selection grabber dots are embedded as
subviews of the text interaction container view, rather than UITextSelectionView itself, so hiding
UITextSelectionView does not affect the visibility of these grabber views (UITextRangeView's -startGrabber and
-endGrabber).

To address this, we augment the change made in r251526 to apply to the start and end grabber views as well.

* WebKitTestRunner/ios/PlatformWebViewIOS.mm:
(WTR::PlatformWebView::windowSnapshotImage):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251858 268f45cc-cd09-0410-ab3c-d52691b4dbfc