webkitpy: Build ImageDiff if it is missing https://bugs.webkit.org/show_bug.cgi?id=183422
Reviewed by Alexey Proskuryakov.

ImageDiff is built with a different SDK than the rest of the WebKit
stack, and this frequently causes infrastructure failures where ImageDiff
is missing on testers. To address this, we should automatically build
ImageDiff if it is missing.

* Scripts/webkitpy/port/base.py:
(Port.check_build): Unconditionally build ImageDiff if it is missing.
(Port.check_image_diff): Use _build_path since _path_to_image_diff will
attempt to use a back-up location.
(Port._path_to_image_diff): If the provided path to ImageDiff does not
exist, use the path of the one we built.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252031 268f45cc-cd09-0410-ab3c-d52691b4dbfc