Remove unimplemented property error log in when getting a property value from a computed style https://bugs.webkit.org/show_bug.cgi?id=203943
Reviewed by Simon Fraser.

In r251581, the list of computed CSS properties was changed to be autogenerated from 'CSSProperties.json'.
One issue that arose from this was that `-webkit-text-decorations-in-effect` was no longer part of the
`CSSComputedStyleDeclaration`, which caused a number of editing tests to fail, as that property is used
for holding internal state related to various editing commands (e.g. `document.execCommand(...)`). In
order to get around this, `ComputedStyleExtractor::copyProperties()` was modified to instead copy all CSS
properties, regardless of whether they're part of the set of computed CSS properties. Unfortunately, this
had a side effect of triggering a longstanding `LOG_ERROR` (added in r6178) when attempting to retrieve
the value of certain CSS properties (e.g. `-internal-text-autosizing-status`).

Since this code path can be triggered by JavaScript (e.g. `getComputedStyle(node).getPropertyCSSValue("...")`),
it doesn't add much benefit to have this log, especially since the code itself is already annotated with
comments/FIXMEs for CSS properties that don't have a computed value.

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::ComputedStyleExtractor::valueForPropertyInStyle):
(WebCore::logUnimplementedPropertyID): Deleted.
* css/SVGCSSComputedStyleDeclaration.cpp:
(WebCore::ComputedStyleExtractor::svgPropertyValue):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252176 268f45cc-cd09-0410-ab3c-d52691b4dbfc