Support or16(TrustedImm32, AbsoluteAddress) in the MIPS MacroAssembler https://bugs.webkit.org/show_bug.cgi?id=204126
Reviewed by Mark Lam.

JSTests:

Decreasing the number of iterations of test to avoid timeout into slow
ports.

* stress/regress-57020338.js:

Source/JavaScriptCore:

This is adjusting the implementation of `or16` for MIPS. This required
the addition of `load16` and `store16` for this macro assembler as
well. This is also fixing build issue on ARMv7 `or16` implementation.

* assembler/MacroAssemblerARMv7.h:
(JSC::MacroAssemblerARMv7::or16):
* assembler/MacroAssemblerMIPS.h:
(JSC::MacroAssemblerMIPS::or16):
(JSC::MacroAssemblerMIPS::load16):
(JSC::MacroAssemblerMIPS::store16):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252463 268f45cc-cd09-0410-ab3c-d52691b4dbfc