results.webkit.org: Report JSC tests to the results database https://bugs.webkit.org/show_bug.cgi?id=204091 <rdar://problem/49778900>
Reviewed by Aakash Jain.

Upload results to the specified results database.

DEMONSTRATION OF FAILURE IN EWS BEHAVIOR, DO NOT LAND!

* Scripts/run-javascriptcore-tests:
(runTest): Parse binary output to mark individual tests as passing
or failing.
(uploadConfiguration): Return a hash map representing the configuration
For this specific test run.
(uploadResults): If a report URL is defined, upload test results to that URL.
* Scripts/webkitdirs.pm:
(splitVersionString): splitVersionString is supposed to be a generally used
function, so should not set a global variable.
(determineOSXVersion): Remove newline from version number.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252490 268f45cc-cd09-0410-ab3c-d52691b4dbfc