Add SPI to clear web process cache https://bugs.webkit.org/show_bug.cgi?id=204210
Patch by Benjamin Nham <nham@apple.com> on 2019-11-15
Reviewed by Chris Dumez.

This adds an SPI to clear cached web processes. The intended use case is to add a lukewarm mode to
Mac PLT5 in which web pages load in a new WebProcess and load resources from the disk cache.

* UIProcess/API/Cocoa/WKProcessPool.mm:
(-[WKProcessPool _clearWebProcessCache]):
* UIProcess/API/Cocoa/WKProcessPoolPrivate.h:
(-[WKProcessPool _clearWebProcessCache]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252500 268f45cc-cd09-0410-ab3c-d52691b4dbfc