Flaky JSC test: stress/stack-overflow-in-yarr-byteCompile.js.dfg-eager. https://bugs.webkit.org/show_bug.cgi?id=204405
Reviewed by Alexey Proskuryakov.

JSTests:

The test was timing out because it's slow.  We don't really need to run this test
on all configurations.  So, we'll change it to on run with the default
configuration.  Also declare the test as "//@ slow!".

* stress/stack-overflow-in-yarr-byteCompile.js:

Source/JavaScriptCore:

$vm.allWithStackSize() manipulates the stack in ways that will freak out ASan.
So, add SUPPRESS_ASAN to functionCallWithStackSize() to tell ASan to ignore it.

* tools/JSDollarVM.cpp:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252694 268f45cc-cd09-0410-ab3c-d52691b4dbfc