Web Inspector: removing the blackbox for a specific script doesn't actually remove the blackbox https://bugs.webkit.org/show_bug.cgi?id=204428
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

Previously, when updating the blackbox state of each existing script, we would only tell the
`Debugger` about when scripts should be blackboxed, not when they shouldn't. This means that
when a given script is un-blackboxed, the `Debugger` would never get told about it and would
therefore still defer pauses as if it was blackboxed.

The solution to this is simple; update the blackboxed state of every script, not just those
that should be blackboxed, and tell the `Debugger` about each.

* inspector/agents/InspectorDebuggerAgent.cpp:
(Inspector::InspectorDebuggerAgent::setShouldBlackboxURL):

LayoutTests:

* inspector/debugger/setShouldBlackboxURL.html:
* inspector/debugger/setShouldBlackboxURL-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252730 268f45cc-cd09-0410-ab3c-d52691b4dbfc