Null check callback in NetworkConnectionToWebProcess::didDeliverMessagePortMessages https://bugs.webkit.org/show_bug.cgi?id=204460 <rdar://problem/57348618>
Reviewed by Chris Dumez.

When the network process crashes during a message port connection establishment,
a web process may have an identifier that becomes stale.  In this case, we should not crash again.
No test because this only happens during a rare race condition after a network process crash, but the radar shows records of this happening.

* NetworkProcess/NetworkConnectionToWebProcess.cpp:
(WebKit::NetworkConnectionToWebProcess::didDeliverMessagePortMessages):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252809 268f45cc-cd09-0410-ab3c-d52691b4dbfc