Ensure SpeechSynthesis::cancel() correctly clears m_currentSpeechUtterance https://bugs.webkit.org/show_bug.cgi?id=204429 <rdar://57072704>
Patch by Sunny He <sunny_he@apple.com> on 2019-11-24
Reviewed by Chris Fleizach.

Currently it is possible for the utterance that is in the process of
being spoken to be garbage collected. In this case, the pointer held
by SpeechSynthesis (m_currentSpeechUtterance) won't be properly
cleared on cancel(). This change ensures that cancel() nulls out the
current utterance and SpeechSynthesis takes a Ref for the utterance.

* Modules/speech/SpeechSynthesis.cpp:
(WebCore::SpeechSynthesis::cancel):
* Modules/speech/SpeechSynthesis.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252841 268f45cc-cd09-0410-ab3c-d52691b4dbfc