mozilla-tests.yaml/js1_5/Array/regress-101964.js is frequently failing on JSC EWS bots. https://bugs.webkit.org/show_bug.cgi?id=200789 <rdar://problem/54361916>
Reviewed by Keith Miller.

JSTests:

The prevailing theory is that this test is being pre-empted and not getting the
CPU time it needs to complete.  As a result, the wall clock time period for
running the test exceeds the expected time.  This patch tests this theory by
changing the time measurement to use CPU time instead.

* mozilla/js1_5/Array/regress-101964.js:

Source/JavaScriptCore:

* tools/JSDollarVM.cpp:
(JSC::functionCurrentCPUTime):
(JSC::JSDollarVM::finishCreation):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253008 268f45cc-cd09-0410-ab3c-d52691b4dbfc