The animation of returning to inline from PiP targets wrong destination rectangle https://bugs.webkit.org/show_bug.cgi?id=204819
Patch by Peng Liu <peng.liu6@apple.com> on 2019-12-03
Reviewed by Jer Noble.

We have to force AVPlayerViewController's view (which is hidden when it is returning to inline in another tab) to run layout
after we set its target size and position to make sure all layers inside the view have the correct information.
Otherwise, the return to inline animation will be incorrect.

* platform/ios/VideoFullscreenInterfaceAVKit.mm:
(VideoFullscreenInterfaceAVKit::preparedToReturnToInline):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253081 268f45cc-cd09-0410-ab3c-d52691b4dbfc