[JSC] Wasm init-expr should reject mutable globals https://bugs.webkit.org/show_bug.cgi?id=205191
Reviewed by Mark Lam.

JSTests:

* wasm/js-api/global-error.js:
(assert.throws.new.WebAssembly.Module.bin):
(new.WebAssembly.Module): Deleted.
(assert.throws): Deleted.
(new.Number): Deleted.

Source/JavaScriptCore:

For init-expr, expr must be a constant[1]. Constant expr, which is defined in Wasm spec, requires that, if the expr is GetGlobal,
global's mutability is immutable. Previously our imported globals are always immutable, so we are using ASSERT instead of checking
mutability in WasmSectionParser. But now, we have ability to import mutable globals. We should check mutability when parsing init-expr.
We do not have this check previously, which leads to spec-correctness issue that we can initialize globals/elements/data-segments
with snapshot values of mutable globals (this is safe, but this is not spec-compliant, and it is not reasonable semantics), while
such an attempt should be rejected when compiling Wasm modules.

This patch adds necessary checks.

[1]: https://webassembly.github.io/spec/core/valid/instructions.html#valid-constant

* wasm/WasmSectionParser.cpp:
(JSC::Wasm::SectionParser::parseInitExpr):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253460 268f45cc-cd09-0410-ab3c-d52691b4dbfc