Behavior of [[GetOwnProperty]] for cross-origin windows is not spec-compliant https://bugs.webkit.org/show_bug.cgi?id=205184
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Rebaseline WPT test now that more checks are passing.

* web-platform-tests/html/browsers/origin/cross-origin-objects/cross-origin-objects-expected.txt:

Source/WebCore:

Behavior of [[GetOwnProperty]] for cross-origin windows is not spec-compliant:
- https://html.spec.whatwg.org/#crossorigingetownpropertyhelper-(-o,-p-)

We should be able to return frames by name, even if their name conflict with the name of a
same-origin window property (e.g. "close"). Previously, we would throw a SecurityError in
this case.

No new tests, rebaselined existing test.

* bindings/js/JSDOMWindowCustom.cpp:
(WebCore::jsDOMWindowGetOwnPropertySlotRestrictedAccess):

LayoutTests:

* http/tests/security/document-all-expected.txt:
The test is accessing the "alert" property on a cross-origin window. We used to throw a SecurityError,
but we now return a Window object since there is a Window whose name is "alert". The test still passes
as it is not able to call alert() cross-origin.

* http/tests/security/xss-DENIED-window-name-navigator-expected.txt:
* http/tests/security/xss-DENIED-window-name-navigator.html:
Update test to use console.log() to print the result instead of alert() since alert() is not allowed
in a sandbox iframe. I updated the expectation so that top.navigator returns the window with the
name "navigator" instead of undefined, as per the behavior change in this patch. I have verified that
our behavior on this test is consistent with Firefox and Chrome.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253483 268f45cc-cd09-0410-ab3c-d52691b4dbfc