MacCatalyst build of libANGLE is installed in the incorrect location https://bugs.webkit.org/show_bug.cgi?id=205219 <rdar://problem/57713353>
Reviewed by Simon Fraser.

Source/ThirdParty/ANGLE:

Rename libANGLE.dylib into libANGLE-shared.dylib, so we can
avoid accidentally trying to link to the libANGLE.a that was
removed recently (but still exists in the SDK for now).

Also, make sure that we install into the correct location for
a Catalyst build.

* ANGLE.xcodeproj/project.pbxproj:
* Configurations/ANGLE.xcconfig:
* Configurations/Base.xcconfig:

Source/WebCore:

The location that WebCore was looking for embedded libraries was
incorrect for Catalyst builds. We never noticed because until now
there were no embedded libraries, and local builds all go into
the same location so this would only happen on Production builds.

Also, libANGLE became libANGLE-shared.

* Configurations/WebCore.xcconfig: Link with libANGLE's new name, and
look in the correct directory.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253499 268f45cc-cd09-0410-ab3c-d52691b4dbfc