Animations stop if new tab opened (and closed) https://bugs.webkit.org/show_bug.cgi?id=202360 <rdar://problem/55923261>
Patch by Antoine Quint <graouts@apple.com> on 2019-12-18
Reviewed by Dean Jackson.

Source/WebCore:

In the case where we would have a fill-forwards software animation when an animation that could be
accelerated started, we would fail to start an accelerated animation because we had no composited
renderer. However, we would still advertise a state of "running accelerated" and the DocumentTimeline
would not schedule ticks to run the animation in software.

We now only schedule accelerated animations once their delay phase is over and the animation is in its
"active" phase, which helps to only schedule accelerated animations once they actually have an effect
that is worth accelerating, and reset pending accelerated actions in case we try to start an accelerated
animation but fail to because we don't have a composited renderer.

Test: webanimations/animation-of-accelerated-property-after-non-accelerated-property.html

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::apply):
(WebCore::KeyframeEffect::updateAcceleratedAnimationState):
(WebCore::KeyframeEffect::applyPendingAcceleratedActions):
* animation/KeyframeEffect.h:

LayoutTests:

Add a ref test that checks that an element with a software animation followed by an animation of a property
that can be accelerated yields animations in both cases.

Also, somehow, the WPT test dom/events/Event-dispatch-on-disabled-elements.html seems to now pass an extra
test that would fail on iOS and now behaves like macOS, so we remove that platform-specific expectation.

* platform/ios/imported/w3c/web-platform-tests/dom/events/Event-dispatch-on-disabled-elements-expected.txt
* webanimations/animation-of-accelerated-property-after-non-accelerated-property-expected.html: Added.
* webanimations/animation-of-accelerated-property-after-non-accelerated-property.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253728 268f45cc-cd09-0410-ab3c-d52691b4dbfc