[CoordinatedGraphics] ThreadedDisplayRefreshMonitor is never released https://bugs.webkit.org/show_bug.cgi?id=205387
Reviewed by Žan Doberšek.

Source/WebCore:

The problem is that DisplayRefreshMonitorManager::createMonitorForClient() always creates a new one for
RenderingUpdateScheduler because it's not notified of the window screen change. So,
createDisplayRefreshMonitor() is called every time, which returns a reference of the same object, but it's added
to the monitors vector of DisplayRefreshMonitorManager and never removed from there.

* page/Chrome.cpp:
(WebCore::Chrome::windowScreenDidChange): Notify the RenderingUpdateScheduler about the screen change.
* page/RenderingUpdateScheduler.h: Make windowScreenDidChange public.

Source/WebKit:

Update the window screen ID after creating the layer tree host to ensure that the call to
createDisplayRefreshMonitor will create the ThreadedDisplayRefrershMonitor instead of the default one.

* WebProcess/WebPage/CoordinatedGraphics/DrawingAreaCoordinatedGraphics.cpp:
(WebKit::DrawingAreaCoordinatedGraphics::enterAcceleratedCompositingMode): Use the dispalyID of the layer tree host.
* WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.cpp:
(WebKit::LayerTreeHost::LayerTreeHost): Initialize the display ID.
* WebProcess/WebPage/CoordinatedGraphics/LayerTreeHost.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253744 268f45cc-cd09-0410-ab3c-d52691b4dbfc