Stop blocking the worker thread in WorkerMessagePortChannelProvider::postMessageToRemote() https://bugs.webkit.org/show_bug.cgi?id=205414
Reviewed by Youenn Fablet.

Source/WebCore:

Stop blocking the worker thread in WorkerMessagePortChannelProvider::postMessageToRemote() as it does not appear
to be needed and it badly impacts performance. This basically replaces a callOnMainThreadAndWait
call (which was added in r249378 as part of a refactoring) with a callOnMainThread call.

This makes fast/workers/worker-cloneport.html runs twice as fast on my machine, which is important
because this test is so slow it is timing out in some configurations.

* dom/MessagePort.cpp:
(WebCore::MessagePort::postMessage):
* dom/messageports/MessagePortChannelProvider.h:
* dom/messageports/MessagePortChannelProviderImpl.cpp:
(WebCore::MessagePortChannelProviderImpl::postMessageToRemote):
* dom/messageports/MessagePortChannelProviderImpl.h:
* dom/messageports/WorkerMessagePortChannelProvider.cpp:
(WebCore::WorkerMessagePortChannelProvider::postMessageToRemote):
* dom/messageports/WorkerMessagePortChannelProvider.h:

Source/WebKit:

* WebProcess/WebCoreSupport/WebMessagePortChannelProvider.cpp:
(WebKit::WebMessagePortChannelProvider::postMessageToRemote):
* WebProcess/WebCoreSupport/WebMessagePortChannelProvider.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253751 268f45cc-cd09-0410-ab3c-d52691b4dbfc