REGRESSION (r253807): crash in storage/indexeddb/modern/opendatabase-request-private.html https://bugs.webkit.org/show_bug.cgi?id=205515
Reviewed by Alex Christensen.

When m_openRequests of IDBTransaction is empty, we expect all requests associated with the transactions should
be completed, but in IDBOpenDBRequest, we removed the request from m_openRequests before
m_currentlyCompletingRequest finished. This is because the order of calling ActiveDOMObject::stop() is random.

* Modules/indexeddb/IDBOpenDBRequest.cpp:
(WebCore::IDBOpenDBRequest::cancelForStop):
* Modules/indexeddb/IDBTransaction.cpp:
(WebCore::IDBTransaction::abortOnServerAndCancelRequests):
(WebCore::IDBTransaction::connectionClosedFromServer):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253858 268f45cc-cd09-0410-ab3c-d52691b4dbfc