REGRESSION (r253282): Tests that use applyAutocorrection assert in UIScriptContext::requestUIScriptCompletion https://bugs.webkit.org/show_bug.cgi?id=205588 <rdar://problem/58109942>
Reviewed by Tim Horton.

The change made in r253282 intended to defer completing the async task prepared in
UIScriptController::applyAutocorrection until after the current runloop; however, it ended up keeping the
synchronous call to asyncTaskComplete. Fix this by removing this code, so that we instead wait for the
dispatch_async block to invoke asyncTaskComplete.

* WebKitTestRunner/ios/UIScriptControllerIOS.mm:
(WTR::UIScriptControllerIOS::applyAutocorrection):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253908 268f45cc-cd09-0410-ab3c-d52691b4dbfc