Declare some classes as final. https://bugs.webkit.org/show_bug.cgi?id=205670
Reviewed by Sam Weinig.

There are a few "Status" classes, all of whom have static computeFor() methods.
All of these classes do not have derived classes, and are independent of each
others in terms of inheritance relationships.  By explicitly declaring them as
final, we can make it clear that a call to any unqualified computeFor() methods
within one of these classes must be from the self class, and that external calls
to any given computeFor() method qualified with a class name is defined in that
class (and is not inherited from another class).

This detail may already be known to folks who are familiar with these classes.
Declaring them as final helps surface this independence for readers of the code
who is not already in the know.

* bytecode/CallLinkStatus.h:
* bytecode/ComplexGetStatus.h:
* bytecode/GetByStatus.h:
* bytecode/InByIdStatus.h:
* bytecode/InstanceOfStatus.h:
* bytecode/PutByIdStatus.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@253969 268f45cc-cd09-0410-ab3c-d52691b4dbfc