highlight/highlight-world-leak.html frequently times out https://bugs.webkit.org/show_bug.cgi?id=205643 <rdar://problem/58263350>
Reviewed by Megan Gardner.

The main logic in this test is executed inside a load event handler to a subframe, which is set in a script
element. However, in the case where the subframe finishes loading before we execute the inline script, the test
will never run, since the load event handler won't be called again.

Fix this by waiting for the load event of the page instead.

* highlight/highlight-world-leak.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254008 268f45cc-cd09-0410-ab3c-d52691b4dbfc